handle invalidId error & some refactoring
