+++ backend/index.js
@@ -1,7 +1,5 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const signUpRoute = require('./routes/auth/sign-up');
-const loginRoute = require('./routes/auth/login');
 const path = require('path');
 const app = express();
 
@@ -33,13 +31,16 @@ app.use(function (req, res, next) {
 /**
  * Auth routes
  */
+const signUpRoute = require('./routes/auth/sign-up');
+const loginRoute = require('./routes/auth/login');
+
 app.use(signUpRoute);
 app.use(loginRoute);
 
 /**
  * Text routes
  */
-const textsRoute = require('./routes/texts/texts');
+const textsRoute = require('./routes/texts');
 
 app.use(textsRoute);
