+++ backend/routes/texts/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Text = require('../../models/text');
+const mongoose = require('mongoose');
 
 router.get('/texts', async (req, res) => {
   try {
@@ -11,16 +12,17 @@ router.get('/texts', async (req, res) => {
     res.json({ error });
   }
 });
+
 router.get('/texts/:textId', async (req, res) => {
   try {
-    const textId = req.params.textId;
-
-    const text = await Text.findOne({ _id: textId });
-
-    if (!text) return res.json({ error: 'No text with such textID' });
+    const text = await Text.findById(req.params.textId);
 
     res.json({ data: text });
   } catch (error) {
+    if (error instanceof mongoose.CastError) {
+      res.json({ error: 'Invalid id' });
+    }
+
     res.json({ error });
   }
 });
